feat(chat,web): docked sharepic artifact panel (Sharepic-Modus)
The variant marked 'Im Chat bearbeiten' now pins as a right-rail artifact
panel on the chat page (>=xl), so it stays in view while the user iterates
via chat instead of scrolling away with the card. Card and panel share the
full render/version/restore logic through the new useSharepicArtifact hook
and the same sharepicLiveStore entry, so SSE updates reach both.

ActiveSharepic carries initialProps/label for panel rendering (client-only;
the request body still maps only variantId/canvasId/canvasType). Also adds
documentation/docs/intern/chat-tool-loop-plan.md — the concrete, code-
verified follow-up plan for the agentic tool loop, scoped v1 to this mode.
