+++ packages/chat/src/components/message-parts/SharepicVariantCard.tsx
@@ -1,4 +1,4 @@
-import { useState, useCallback, useEffect, useRef } from 'react';
+import { useCallback } from 'react';
 import {
   Loader2,
   Pencil,
@@ -10,8 +10,7 @@ import {
   History,
 } from 'lucide-react';
 import { cn } from '../../lib/utils';
-import { useChatConfigStore } from '../../stores/chatConfigStore';
-import { useSharepicLiveStore } from '../../stores/sharepicLiveStore';
+import { useSharepicArtifact } from '../../hooks/useSharepicArtifact';
 
 import type { SharepicVariant } from '../../hooks/useChatGraphStream';
 
@@ -19,222 +18,36 @@ interface SharepicVariantCardProps {
   variant: SharepicVariant;
 }
 
-const FALLBACK_LABELS: Record<string, string> = {
-  dreizeilen: 'Dreizeiler',
-  'zitat-pure': 'Zitat',
-  zitat: 'Zitat',
-  info: 'Info',
-  simple: 'Sharepic',
-  veranstaltung: 'Veranstaltung',
-  slider: 'Slider',
-  freeform: 'Freeform',
-};
-
-interface VersionEntry {
-  version: number;
-  summary: string | null;
-}
-
-/**
- * After a real edit (entry is thumbnail-dirty) the freshly rendered head PNG
- * doubles as the canvas thumbnail. Version previews never qualify, and the
- * flag is cleared up front — a failed upload is not worth a retry loop.
- */
-function maybeUploadThumbnail(variantId: string, dataUrl: string, isVersionPreview: boolean) {
-  if (isVersionPreview) return;
-  const store = useSharepicLiveStore.getState();
-  const entry = store.entries[variantId];
-  if (!entry?.thumbnailDirty || !entry.canvasId) return;
-  const upload = useChatConfigStore.getState().updateSharepicThumbnail;
-  if (!upload) return;
-  store.clearThumbnailDirty(variantId);
-  upload(entry.canvasId, dataUrl).catch((err) => {
-    console.warn('[SharepicVariantCard] Thumbnail-Update fehlgeschlagen:', err);
-  });
-}
-
 export function SharepicVariantCard({ variant }: SharepicVariantCardProps) {
-  const [imageBase64, setImageBase64] = useState<string | null>(null);
-  const [isRendering, setIsRendering] = useState(true);
-  const [renderError, setRenderError] = useState(false);
-  const [versions, setVersions] = useState<VersionEntry[] | null>(null);
-  /** Version being previewed via the stepper; null = current head. */
-  const [viewVersion, setViewVersion] = useState<number | null>(null);
-  const [viewState, setViewState] = useState<Record<string, unknown> | null>(null);
-  const versionStateCache = useRef(new Map<number, Record<string, unknown>>());
-  const hydratedRef = useRef(false);
-
-  const live = useSharepicLiveStore((s) => s.entries[variant.id]);
-  const isActiveForChat = useSharepicLiveStore((s) => s.activeVariant?.variantId === variant.id);
-
-  const canvasId = live?.canvasId ?? variant.canvasId ?? null;
-  const headVersion = live?.version ?? null;
-  const renderInput = viewState ?? live?.state ?? variant.initialProps;
-
-  // Render (and re-render after each chat edit / version step). renderInput
-  // is the full flat state — StandaloneCanvas's createInitialState accepts it
-  // in place of the original initialProps.
-  useEffect(() => {
-    let cancelled = false;
-    const renderFn = useChatConfigStore.getState().renderSharepic;
-    if (!renderFn) {
-      setRenderError(true);
-      setIsRendering(false);
-      return undefined;
-    }
-    setIsRendering(true);
-    renderFn(variant.canvasType, renderInput)
-      .then((dataUrl) => {
-        if (cancelled) return;
-        if (dataUrl) {
-          setImageBase64(dataUrl);
-          setRenderError(false);
-          maybeUploadThumbnail(variant.id, dataUrl, viewState != null);
-        } else {
-          setRenderError(true);
-        }
-      })
-      .catch(() => {
-        if (!cancelled) setRenderError(true);
-      })
-      .finally(() => {
-        if (!cancelled) setIsRendering(false);
-      });
-    return () => {
-      cancelled = true;
-    };
-  }, [variant.canvasType, variant.id, renderInput, viewState]);
-
-  // Thread-reload rehydration: a minted variant renders its CURRENT state
-  // (which may have changed in the studio), not the stale initialProps.
-  useEffect(() => {
-    if (!canvasId || hydratedRef.current || live?.state) return;
-    const fetchState = useChatConfigStore.getState().fetchSharepicState;
-    if (!fetchState) return;
-    hydratedRef.current = true;
-    void fetchState(canvasId).then((result) => {
-      if (!result) return;
-      useSharepicLiveStore.getState().upsertEntry(variant.id, {
-        canvasId,
-        canvasType: variant.canvasType,
-        version: result.version,
-        state: result.state,
-      });
-    });
-  }, [canvasId, live?.state, variant.id, variant.canvasType]);
-
-  // New head version (chat edit applied) → drop any stale version preview.
-  useEffect(() => {
-    setViewVersion(null);
-    setViewState(null);
-    setVersions(null);
-  }, [headVersion]);
-
-  const loadVersions = useCallback(async (): Promise<VersionEntry[]> => {
-    if (versions) return versions;
-    const fetchVersions = useChatConfigStore.getState().fetchSharepicVersions;
-    if (!canvasId || !fetchVersions) return [];
-    const list = await fetchVersions(canvasId).catch(() => []);
-    const entries = list.map((v) => ({ version: v.version, summary: v.summary }));
-    setVersions(entries);
-    return entries;
-  }, [canvasId, versions]);
-
-  const stepToVersion = useCallback(
-    async (direction: -1 | 1) => {
-      if (!canvasId) return;
-      const list = await loadVersions();
-      if (list.length < 2) return;
-      const ordered = [...list].sort((a, b) => a.version - b.version);
-      const currentIdx =
-        viewVersion == null
-          ? ordered.length - 1
-          : ordered.findIndex((v) => v.version === viewVersion);
-      const nextIdx = Math.min(Math.max(currentIdx + direction, 0), ordered.length - 1);
-      const target = ordered[nextIdx];
-      if (!target) return;
-      if (nextIdx === ordered.length - 1) {
-        setViewVersion(null);
-        setViewState(null);
-        return;
-      }
-      const cached = versionStateCache.current.get(target.version);
-      if (cached) {
-        setViewVersion(target.version);
-        setViewState(cached);
-        return;
-      }
-      const fetchVersionState = useChatConfigStore.getState().fetchSharepicVersionState;
-      if (!fetchVersionState) return;
-      const state = await fetchVersionState(canvasId, target.version).catch(() => null);
-      if (state) {
-        versionStateCache.current.set(target.version, state);
-        setViewVersion(target.version);
-        setViewState(state);
-      }
-    },
-    [canvasId, loadVersions, viewVersion]
-  );
-
-  const handleRestore = useCallback(async () => {
-    if (!canvasId || viewVersion == null) return;
-    const restore = useChatConfigStore.getState().restoreSharepicVersion;
-    if (!restore) return;
-    const result = await restore(canvasId, viewVersion).catch(() => null);
-    if (result) {
-      useSharepicLiveStore.getState().upsertEntry(variant.id, {
-        canvasId,
-        canvasType: variant.canvasType,
-        version: result.version,
-        state: result.state,
-        thumbnailDirty: true,
-      });
-    }
-  }, [canvasId, viewVersion, variant.id, variant.canvasType]);
-
-  const handleToggleActive = useCallback(() => {
-    const store = useSharepicLiveStore.getState();
-    if (store.activeVariant?.variantId === variant.id) {
-      store.setActiveVariant(null);
-    } else {
-      store.setActiveVariant({
-        variantId: variant.id,
-        canvasId,
-        canvasType: variant.canvasType,
-      });
-    }
-  }, [variant.id, variant.canvasType, canvasId]);
+  const {
+    imageBase64,
+    isRendering,
+    renderError,
+    headVersion,
+    viewVersion,
+    isActiveForChat,
+    showStepper,
+    label,
+    stepToVersion,
+    restoreViewVersion,
+    toggleActive,
+    download,
+    openInStudio,
+  } = useSharepicArtifact(variant);
 
   const handleDownload = useCallback(
     (e: React.MouseEvent) => {
       e.stopPropagation();
-      if (!imageBase64) return;
-      const link = document.createElement('a');
-      link.href = imageBase64;
-      link.download = `sharepic-${variant.canvasType}.png`;
-      document.body.appendChild(link);
-      link.click();
-      document.body.removeChild(link);
+      download();
     },
-    [imageBase64, variant.canvasType]
+    [download]
   );
-
-  const handleEdit = useCallback(() => {
-    useChatConfigStore.getState().onEditSharepic?.({
-      ...variant,
-      initialProps: live?.state ?? variant.initialProps,
-      ...(canvasId ? { canvasId } : {}),
-    });
-  }, [variant, live?.state, canvasId]);
-
-  const label = variant.label ?? FALLBACK_LABELS[variant.canvasType] ?? 'Sharepic';
-  const showStepper = canvasId != null && headVersion != null && headVersion > 1;
 
   if (renderError) {
     return (
       <div className="rounded-lg border border-border p-4 text-sm text-foreground-muted">
         Sharepic-Vorschau konnte nicht gerendert werden.
-        <button onClick={handleEdit} className="ml-2 text-primary hover:underline">
+        <button onClick={openInStudio} className="ml-2 text-primary hover:underline">
           Im Editor öffnen
         </button>
       </div>
@@ -252,7 +65,7 @@ export function SharepicVariantCard({ variant }: SharepicVariantCardProps) {
     >
       <button
         type="button"
-        onClick={handleEdit}
+        onClick={openInStudio}
         className="block w-full text-left focus:outline-none focus-visible:ring-2 focus-visible:ring-primary"
         aria-label={`${label}-Variante im Editor öffnen`}
       >
@@ -319,7 +132,7 @@ export function SharepicVariantCard({ variant }: SharepicVariantCardProps) {
             )}
             {viewVersion != null && (
               <button
-                onClick={() => void handleRestore()}
+                onClick={() => void restoreViewVersion()}
                 className="flex items-center gap-1 rounded-lg px-2 py-1 text-xs text-primary hover:bg-primary/10"
                 aria-label={`Version ${viewVersion} wiederherstellen`}
               >
@@ -330,7 +143,7 @@ export function SharepicVariantCard({ variant }: SharepicVariantCardProps) {
           </div>
           <div className="flex items-center gap-1">
             <button
-              onClick={handleToggleActive}
+              onClick={toggleActive}
               className={cn(
                 'flex items-center gap-1 rounded-lg px-2 py-1 text-xs',
                 isActiveForChat
@@ -352,7 +165,7 @@ export function SharepicVariantCard({ variant }: SharepicVariantCardProps) {
               <span>Herunterladen</span>
             </button>
             <button
-              onClick={handleEdit}
+              onClick={openInStudio}
               className="flex items-center gap-1 rounded-lg px-2 py-1 text-xs text-primary hover:bg-primary/10"
               aria-label="Sharepic im Studio bearbeiten"
             >
